+++ app.js
@@ -4,9 +4,8 @@ const cors = require('cors');
 const globalErrorHandler = require('./middlewares/globalErrorHandler');
 const notFoundHandler = require('./middlewares/notFoundHandler');
 require('dotenv').config();
-
 const contactsRouter = require('./routes/api/contacts');
-const authRouter = require('./routes/api/auth');
+const usersRouter = require('./routes/api/users');
 
 const app = express();
 
@@ -19,7 +18,7 @@ app.use(cors());
 app.use(express.json());
 
 app.use('/api/contacts', contactsRouter);
-app.use('/api/auth', authRouter);
+app.use('/api/users', usersRouter);
 
 app.use(notFoundHandler);
