+++ routes/api/users.js
@@ -5,14 +5,14 @@ const {
   loginController,
   getCurrentController,
   logoutController,
-} = require('../../controllers/auth');
+  updateSubscriptionController,
+} = require('../../controllers/users');
 
 const { validateBody, authenticate } = require('../../middlewares');
 
-const { registerSchema, loginSchema } = require('../../schemas');
+const { registerSchema, loginSchema, updateSubscriptionSchema } = require('../../schemas');
 
 const router = express.Router();
-
 
 router.post('/register', validateBody(registerSchema), registerController);
 
@@ -21,5 +21,7 @@ router.post('/login', validateBody(loginSchema), loginController);
 router.get('/current', authenticate, getCurrentController);
 
 router.post('/logout', authenticate, logoutController);
+
+router.patch('/', authenticate, validateBody(updateSubscriptionSchema), updateSubscriptionController);
 
 module.exports = router;
