+++ schemas/usersSchema.js
@@ -1,8 +1,7 @@
 const Joi = require('joi');
-const { emailRegexp } = require('../utils');
+const { emailRegexp, subscriptionList } = require('../utils');
 
 const registerSchema = Joi.object({
-//   name: Joi.string().required(),
   email: Joi.string().pattern(emailRegexp).required(),
   password: Joi.string().min(6).required(),
 });
@@ -12,12 +11,14 @@ const loginSchema = Joi.object({
   password: Joi.string().min(6).required(),
 });
 
-// const schemas = {
-//   registerSchema,
-//   loginSchema,
-// };
+const updateSubscriptionSchema = Joi.object({
+  subscription: Joi.string()
+    .valid(...subscriptionList)
+    .required(),
+});
 
 module.exports = {
   registerSchema,
   loginSchema,
+  updateSubscriptionSchema,
 };
