+++ utils/index.js
@@ -1,7 +1,9 @@
 const HttpError = require('./HttpError');
 const { emailRegexp } = require('./regexps');
+const { subscriptionList } = require('./subscriptionList');
 
 module.exports = {
   HttpError,
   emailRegexp,
+  subscriptionList,
 };
