+++ utils/regexps.js
@@ -1,6 +1,5 @@
 const emailRegexp = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/;
 
-
 module.exports = {
   emailRegexp,
 };
