+++ utils/subscriptionList.js
@@ -1,0 +1,5 @@
+const subscriptionList = ['starter', 'pro', 'business'];
+
+module.exports = {
+  subscriptionList,
+};
